api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.35
psycopg2-binary==2.9.9
pydantic==2.9.2
pydantic-settings==2.4.0
python-jose==3.3.0
argon2-cffi==23.1.0
passlib==1.7.4
python-multipart==0.0.9
pandas==2.2.2
openpyxl==3.1.5
